requirements.txt
asgiref==3.8.1
certifi==2019.9.11
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
commonmark==0.9.1
cryptography==43.0.0
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==5.0.7
django-cors-headers==4.4.0
django-import-export==4.1.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
geographiclib==2.0
geopy==2.4.1
idna==3.7
imageio==2.34.2
import-export==0.3.1
mysqlclient==2.2.4
numpy==2.0.1
oauthlib==3.2.2
opencv-python==4.10.0.84
pillow==10.4.0
pip==24.2
py-localtunnel==1.0.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python3-openid==3.2.0
rawpy==0.22.0
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.14.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==9.1